% Generated for baobabStats 1.1.0
\name{bs_geo_cameroun}
\alias{bs_geo_cameroun}
\title{Referentiel geographique hierarchique du Cameroun}
\usage{
bs_geo_cameroun(niveau = c("complet", "region", "departement", "arrondissement"))
}
\arguments{
\item{niveau}{Niveau de detail souhaite.}
}
\value{Un tibble des unites administratives au niveau demande.}
\description{
Repertoire officiel des localites du Cameroun (repertoire des localites et
villages 2016) : 10 regions, 58 departements, 360 arrondissements, plus de
1 600 villes/cantons. Base de l'harmonisation geographique et de la jointure
des cartes thematiques.
}
